sap_erp/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
autopep8==2.3.2
bedrock-agentcore==1.0.5
bedrock-agentcore-starter-toolkit==0.1.28
boto3==1.40.65
botocore==1.40.65
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
docstring_parser==0.17.0
fastapi==0.121.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.20.0
mdurl==0.1.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pathable==0.4.4
prance==25.4.8.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-openapi-schema-to-json-schema==0.0.3
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
questionary==2.1.1
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.49.3
strands-agents==1.14.0
toml==0.10.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
wrapt==1.17.3
zipp==3.23.0
